git-svn-id: https://gorgonlib.googlecode.com/svn/branches/2.x@1790 a61f9e97-bd4c-0410-8c65-85b29eaa70ff

+++ Dorian/PlugIns/Gorgon.Input.RawInput/Native/Win32Enums.cs
@@ -611,7 +611,7 @@ namespace GorgonLibrary.Native
 	/// </summary>
 	[Flags]
 	enum RawMouseFlags
-		: ushort
+		: uint
 	{
 		/// <summary>Relative to the last position.</summary>
 		MoveRelative = 0,
